test: 입력 유효성 검사 테스트 코드 수정 및 컴퓨터 랜덤 숫자 테스트 코드 작성
- 입력 유효성 검사하는 테스트 코드 분리 (userInputCheck.test.js)
- checkUserInput 함수에 대해 독립적이고 실행속도도 더 빠른 테스트 코드로 수정
- 컴퓨터 난수 생성하는 메서드에 대해 실패하는 테스트 코드 작성 (App.test.js)
